Add Invoice download in one click, filename structure.
Added Functionality of downloading all the invoices in one click,
modified the filename structure as "invoicenumber_consumername". The
invoice will be downloaded as png and when dowmloading all of them in
one click it will be in zip file. Also added log messaged to debug when
there is no consumer name or invoice number in certain cells.